requirements.txt
flask==3.1.0
flask-cors==6.0.1
python-dotenv==1.1.1
pyodbc==5.2.0